third_party/blink/web_tests/external/wpt/clipboard-apis/clipboard-item.https-expected.txt

This is a testharness.js-based test.
[FAIL] supports(text/uri-list) returns true
  assert_equals: expected true but got false
Harness: the test ran to completion.
